fix paddle bug

+++ paddlepaddle/paddlepaddle-engine/src/main/java/ai/djl/paddlepaddle/engine/PpDataType.java
@@ -27,8 +27,8 @@ public final class PpDataType {
     private static Map<DataType, Integer> createMapToPaddlePaddle() {
         Map<DataType, Integer> map = new ConcurrentHashMap<>();
         map.put(DataType.FLOAT32, 0);
-        map.put(DataType.INT32, 1);
-        map.put(DataType.INT64, 2);
+        map.put(DataType.INT64, 1);
+        map.put(DataType.INT32, 2);
         map.put(DataType.INT8, 3);
         map.put(DataType.UINT8, 3);
         map.put(DataType.UNKNOWN, 4);
@@ -38,8 +38,8 @@ public final class PpDataType {
     private static Map<Integer, DataType> createMapFromPaddlePaddle() {
         Map<Integer, DataType> map = new ConcurrentHashMap<>();
         map.put(0, DataType.FLOAT32);
-        map.put(1, DataType.INT32);
-        map.put(2, DataType.INT64);
+        map.put(1, DataType.INT64);
+        map.put(2, DataType.INT32);
         map.put(3, DataType.INT8);
         map.put(4, DataType.UNKNOWN);
         return map;
